* master: (38 commits)
  tests: fix missing dependency of thermodynamics on clock
  connect runtime estimator to timer_service
  add separate simulation clock, mdsim::clock
  use simulation step as time step (instead of time)
  init.lua: instantiate halmd.random in every case
  gpu::verlet: remove unused class attribute 'device'
  device: explicitly write to std::cerr in the destructor
  device: add debug message in destructor
  {positions,velocities}::phase_space: add assertions
  box: read parameter 'length' upon continuation
  rand48_kernel: fix FTBFS with CUDA 4.0 (RC2).
  Revert "doc: configure Sphinx extension breathe."
  Revert "doc: generate reference using breathe."
  Revert "gpu/density_mode: render formula using inline ReST."
  gpu/density_mode: render formula using inline ReST.
  doc: define doxygen aliases \rst and \endrst.
  doc: generate reference using breathe.
  doc: configure Sphinx extension breathe.
  Add periodic timer service.
  Revert "doc: disable LaTeX compiler during doxygen run."
  ...

Conflicts:
	test/mdsim/integrators/CMakeLists.txt

Additional actions:
	test/unit/mdsim/integrators/verlet_nvt_hoover.cpp

+++ halmd/observables/phase_space.cpp
@@ -32,7 +32,7 @@ namespace observables
 {
 
 template <typename phase_space_type>
-typename signal<void (double)>::slot_function_type
+typename signal<void (uint64_t)>::slot_function_type
 acquire_wrapper(shared_ptr<phase_space_type> phase_space)
 {
     return bind(&phase_space_type::acquire, phase_space, _1);
